--- apps/web/app/calma pae/admin/page.tsx
@@ -1,9 +1,0 @@
-import SuggestTrailsAdmin from "./components";
-
-export default function AdminSuggestTrailsPage() {
-  return (
-    <div className="w-full min-h-screen bg-slate-50 px-4 py-8">
-      <SuggestTrailsAdmin />
-    </div>
-  );
-}
