--- apps/web/app/calma pae/page.tsx.example
@@ -1,5 +1,0 @@
-import { redirect } from "next/navigation";
-
-export default function SuggestTrailsPage() {
-  redirect("/suggest-trails/hikers");
-}
