+++ apps/web/app/suggest-trails/_components/suggest-trail-whiteboard.tsx
