+++ apps/web/app/suggested-trails/page.tsx
@@ -1,0 +1,9 @@
+import SuggestedTrailsAdmin from "./_components/suggested-trails-admin";
+
+export default function SuggestedTrailsPage() {
+  return (
+    <div className="w-full min-h-screen bg-slate-50 px-4 py-8">
+      <SuggestedTrailsAdmin />
+    </div>
+  );
+}
