+++ apps/web/app/suggested-trails/types.ts
@@ -1,0 +1,13 @@
+export type SuggestionStatus = "novo" | "a fazer" | "em andamento";
+
+export interface TrailSuggestion {
+  id: string;
+  name: string;
+  location: string;
+  lengthKm: string;
+  description: string;
+  status: SuggestionStatus;
+  createdAt: string;
+}
+
+export const SUGGESTED_TRAILS_STORAGE_KEY = "suggested-trails";
